:technologist: [#2478] -- extract base selenium test case
Taken from the existing test, this provides an envvar-configurable
selenium setup for future/extra selenium tests.
