+++ src/openforms/tests/selenium_tests/test_admin_js.py
@@ -1,53 +1,18 @@
 """
 Tests for Javascript bundle running in the admin context.
 """
-import os
-from unittest import skipIf
-
-from django.contrib.staticfiles.testing import StaticLiveServerTestCase
-from django.test import override_settings
 from django.urls import reverse
-from django.utils.module_loading import import_string
 
 from furl import furl
 from selenium.common import NoSuchElementException
 from selenium.webdriver.common.by import By
 
 from openforms.accounts.tests.factories import SuperUserFactory
-from openforms.tests.utils import disable_2fa
 
-SELENIUM_WEBDRIVER = os.getenv("SELENIUM_WEBDRIVER", default="Chrome")
-SELENIUM_HEADLESS = "NO_SELENIUM_HEADLESS" not in os.environ
-
-if SELENIUM_WEBDRIVER:
-    WebDriver = import_string(f"selenium.webdriver.{SELENIUM_WEBDRIVER}")
-    Options = import_string(
-        f"selenium.webdriver.{SELENIUM_WEBDRIVER.lower()}.options.Options"
-    )
-else:
-    WebDriver, Options = None, None
+from .base import SeleniumTestCase
 
 
-@skipIf(not WebDriver, "No Selenium webdriver configured")
-@disable_2fa
-@override_settings(ALLOWED_HOSTS=["*"])
-class GeneralAdminJSTests(StaticLiveServerTestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-
-        options = Options()
-        if SELENIUM_HEADLESS:
-            options.headless = True
-
-        cls.selenium = WebDriver(options=options)
-        cls.selenium.implicitly_wait(3)
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.selenium.quit()
-        super().tearDownClass()
-
+class GeneralAdminJSTests(SeleniumTestCase):
     def test_admin_pages_with_js_bundle_dont_crash(self):
         SuperUserFactory.create(username="admin", password="selenium")
